// src/common/enums/weather-condition.enum.ts
export enum WeatherCondition {
  CLEAR = 'CLEAR',

  PARTLY_CLOUDY = 'PARTLY_CLOUDY',
  OVERCAST = 'OVERCAST',

  FOG = 'FOG',

  DRIZZLE = 'DRIZZLE',
  RAIN = 'RAIN',
  HEAVY_RAIN = 'HEAVY_RAIN',
  FREEZING_RAIN = 'FREEZING_RAIN',

  SNOW = 'SNOW',
  SNOW_SHOWERS = 'SNOW_SHOWERS',

  RAIN_SHOWERS = 'RAIN_SHOWERS',

  THUNDERSTORM = 'THUNDERSTORM',
}